+++ src/app/shared/enums.ts
@@ -71,3 +71,14 @@ export enum WorkflowStep {
   DisplayExtractedText = 'DISPLAY_EXTRACTED_TEXT',
   DisplayResultsFromStorage = 'DISPLAY_RESULTS_FROM_STORAGE',
 }
+
+export enum FileNamePrefix {
+  QRCode = 'qrcode',
+  ImageToText = 'img_txt',
+}
+
+export enum FileExtension {
+  PNG = 'png',
+  JPEG = 'jpeg',
+  PDF = 'pdf',
+}