+++ src/app/ui/components/photo-info-modal/photo-info-modal.component.spec.ts
@@ -1,0 +1,24 @@
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { IonicModule } from '@ionic/angular';
+
+import { PhotoInfoModalComponent } from './photo-info-modal.component';
+
+describe('PhotoInfoModalComponent', () => {
+  let component: PhotoInfoModalComponent;
+  let fixture: ComponentFixture<PhotoInfoModalComponent>;
+
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      declarations: [ PhotoInfoModalComponent ],
+      imports: [IonicModule.forRoot()]
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(PhotoInfoModalComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  }));
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});